backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo[srv]==4.5.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
Werkzeug==2.3.7
gunicorn==21.2.0 